chore: Refactors action to have all PRNumber related logic in the input reading phase
